fixed job posting and cv view to include edit as well as create

+++ app/Http/Controllers/ApplicationController.php
@@ -8,32 +8,12 @@ use Illuminate\Http\Request;
 class ApplicationController extends Controller
 {
 
-    public function __construct()
-    {
-        $this->authorizeResource(Application::class, 'apply');
-    }
-
-    /**
-     * Display a listing of the resource.
-     */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
-    }
-
     /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Application::class);
         $application = new Application();
         $application->user()->associate($request->user());
         $application->job()->associate($request->input('job'));
@@ -41,34 +21,11 @@ class ApplicationController extends Controller
     }
 
     /**
-     * Display the specified resource.
-     */
-    public function show(Application $application)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(Application $application)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, Application $application)
-    {
-        //
-    }
-
-    /**
      * Remove the specified resource from storage.
      */
     public function destroy(Application $application)
     {
+        $this->authorize('delete', $application);
         $application->delete();
     }
 }
